change to int

+++ source/controllers/enemies/ShotgunnerController.cpp
@@ -32,7 +32,7 @@ void ShotgunnerController::attackPlayer(std::shared_ptr<EnemyModel> enemy,
       enemy->addBullet(p);
     }
     if (enemy->getAttackCooldown() <= 0) {
-      std::uniform_int_distribution<float> dist(0.0f, 50.0f);
+      std::uniform_int_distribution<int> dist(0.0f, 50.0f);
       enemy->setAttackCooldown(dist(_generator) + ATTACK_COOLDOWN);
     }
   } else {
